feat: Map (x2) -> add condition

+++ api/src/main/java/fr/plop/generic/position/Point.java
@@ -1,6 +1,6 @@
 package fr.plop.generic.position;
 
-public record Point(float lat, float lng) {
+public record Point(double lat, double lng) {
     public boolean in(Point bottomLeft, Point topRight) {
         return bottomLeft.lat() <= lat && lat <= topRight.lat()
                 && bottomLeft.lng() <= lng && lng <= topRight.lng();
